backend/location_weather/requirements.txt
Flask==3.1.0
Flask_Cors==5.0.0
PyJWT==2.10.1
python-dotenv==1.0.1
Requests==2.32.3
supabase==2.13.0